song.rsを4つのサブモジュールに分割 #75
song.rs (816行) のデータ構造を役割ごとのサブモジュールに分割する。

- song/track.rs    トラックと演奏パラメータの管理 (Track / NoteInfo /
                   ControlChangeOnNoteWave)
- song/event.rs    MIDIイベントの定義 (Event / EventType)
- song/function.rs ユーザー定義関数・システム関数の定義 (SFunction / SFunctionType)
- song/flags.rs    実行時フラグ (Flags)

Song 構造体と各種定数は song.rs に残した。

移動した型は pub のままとし、親モジュールで pub use track::*; のように
再エクスポートしているため、crate::song::Event などの既存のパスは
すべてそのまま利用できる。公開APIは変更していない。

このコミットは移動のみで振る舞いは変更していない。
cargo test 102件通過、ゴールデンMIDIのハッシュも一致。
