+++ src/song.rs
@@ -1,5 +1,16 @@
 //! song & track
 
+mod event;
+mod flags;
+mod function;
+mod track;
+
+pub use event::*;
+pub use flags::*;
+pub use function::*;
+pub use track::*;
+
+
 use std::collections::HashMap;
 use crate::runner::value_range;
 use crate::sakura_functions;
@@ -12,613 +23,6 @@ use crate::token::Tokens;
 pub const SAKURA_MAX_LOGS: usize = 100; // lines
 pub const SAKURA_MAX_LOGS_CHARS: usize = 1024 * 4; // chars
 pub const SAKURA_DEFAULT_RANDOM_SEED: u32 = 3958587042; // random seed
-
-/// Event Type
-#[derive(Debug, Clone, PartialEq)]
-pub enum EventType {
-    NoteOn,
-    NoteOff,
-    ControllChange,
-    PitchBend,
-    PitchBendRange,
-    Voice,
-    Meta,
-    SysEx,
-    DirectSMF,
-}
-
-/// Event
-#[derive(Debug, Clone)]
-pub struct Event {
-    pub etype: EventType,
-    pub time: isize,
-    pub channel: isize,
-    pub v1: isize,
-    pub v2: isize,
-    pub v3: isize,
-    pub data: Option<Vec<u8>>,
-}
-impl Event {
-    pub fn note(time: isize, channel: isize, note_no: isize, len: isize, vel: isize) -> Self {
-        Self { etype: EventType::NoteOn, time, channel, v1: note_no, v2: len, v3: vel, data: None }
-    }
-    pub fn voice(time: isize, channel: isize, value: isize) -> Self {
-        Self { etype: EventType::Voice, time, channel, v1: value, v2: 0, v3: 0, data: None }
-    }
-    pub fn meta(time: isize, v1: isize, v2: isize, v3: isize, data_v: Vec<u8>) -> Self {
-        Self { etype: EventType::Meta, time, channel: 0, v1, v2, v3, data: Some(data_v) }
-    }
-    /// generate SMF event type
-    pub fn direct_smf(time: isize, data_v: Vec<u8>) -> Self {
-        Self { etype: EventType::DirectSMF, time, channel:0, v1: 0 , v2: 0, v3: 0, data: Some(data_v) }
-    }
-    pub fn sysex(time: isize, data_v: &Vec<SValue>, checksum_mode: bool) -> Self {
-        // convert to u8 withaout checksum
-        if checksum_mode == false {
-            let mut a: Vec<u8> = vec![];
-            for v in data_v.iter() {
-                a.push(v.to_i() as u8);
-            }
-            return Self { etype: EventType::SysEx, time, channel: 0, v1: 0, v2: 0, v3: 0, data: Some(a) };
-        }
-        // calc checksum
-        let mut a: Vec<u8> = vec![];
-        let mut checksum: isize = 0;
-        let mut flag_checksum = false;
-        for v in data_v.iter() {
-            let n = v.to_i();
-            if flag_checksum {
-                if n == -2 {
-                    flag_checksum = false;
-                    let checksum_v = ((128 - (checksum & 0x7F)) & 0x7F) as u8; // 7bit補正
-                    a.push(checksum_v);
-                    continue;
-                } else {
-                    checksum += n;
-                }
-            }
-            if n == -1 {
-                flag_checksum = true;
-                continue;
-            }
-            a.push(n as u8);
-        }
-        Self { etype: EventType::SysEx, time, channel: 0, v1: 0, v2: 0, v3: 0, data: Some(a) }
-    }
-    
-    pub fn sysex_raw(time: isize, data_v: Vec<u8>) -> Self {
-        Self { etype: EventType::SysEx, time, channel: 0, v1: 0, v2: 0, v3: 0, data: Some(data_v) }
-    }
-    /// ControllChange
-    pub fn cc(time: isize, channel: isize, no: isize, value: isize) -> Self {
-        Self { etype: EventType::ControllChange, time, channel, v1: no, v2: value, v3:0, data: None }
-    }
-    /// pitch_bend : 0..16383 (-8192 .. 0 .. 8191)
-    pub fn pitch_bend(time: isize, channel: isize, value: isize) -> Self {
-        Self { etype: EventType::PitchBend, time, channel, v1: value, v2: 0, v3: 0, data: None }
-    }
-    pub fn pitch_bend_range(time: isize, channel: isize, value: isize) -> Self {
-        Self { etype: EventType::PitchBendRange, time, channel, v1: value, v2: 0, v3: 0, data: None }
-    }
-    /// dump data
-    pub fn dump_data_to_hexstr(&self) -> String {
-        let mut r = vec![];
-        match &self.data {
-            None => return "".to_string(),
-            Some(data) => {
-                for v in data.iter() {
-                    r.push(format!("{:02X}", v));
-                }
-            }
-        }
-        r.join(",")
-    }
-}
-
-/// NoteInfo
-#[derive(Debug)]
-pub struct NoteInfo {
-    pub no: isize,
-    pub flag: isize,
-    pub natural: isize,
-    pub len_s: String,
-    pub qlen: isize,
-    pub vel: isize,
-    pub t: isize,
-    pub o: isize,
-    pub slur: isize,
-}
-
-
-#[derive(Debug, Clone)]
-pub struct ControlChangeOnNoteWave {
-    pub no: isize,
-    pub data: Vec<isize>,
-    pub index: isize, // for ControlChangeOnNote
-}
-/// Track
-#[derive(Debug)]
-pub struct Track {
-    pub timepos: isize,
-    pub channel: isize,
-    pub length: isize,
-    pub octave: isize,
-    pub velocity: isize,
-    pub v_sub: Vec<isize>,
-    pub qlen: isize,
-    pub timing: isize,
-    pub v_rand: isize,
-    pub q_rand: isize,
-    pub t_rand: isize,
-    pub o_rand: isize,
-    pub port: isize,
-    pub track_key: isize,
-    pub tie_mode: TieMode, // Slur(#7)
-    pub tie_value: isize,
-    pub bend_range: isize,
-    pub program_change: isize,
-    pub v_on_time_start: isize,
-    pub v_on_time: Option<Vec<isize>>,
-    pub v_on_note_index: isize,
-    pub v_on_note_is_cycle: bool,
-    pub v_on_note: Option<Vec<isize>>,
-    pub q_on_note_index: isize,
-    pub q_on_note: Option<Vec<isize>>,
-    pub q_on_note_is_cycle: bool,
-    pub t_on_note_index: isize,
-    pub t_on_note: Option<Vec<isize>>,
-    pub t_on_note_is_cycle: bool,
-    pub o_on_note_index: isize,
-    pub o_on_note: Option<Vec<isize>>,
-    pub o_on_note_is_cycle: bool,
-    pub l_on_note_index: isize,
-    pub l_on_note: Option<Vec<isize>>,
-    pub l_on_note_is_cycle: bool,
-    pub cc_on_time_freq: isize,
-    pub events: Vec<Event>,
-    pub tie_notes: Vec<Event>,
-    pub cc_on_note: Vec<ControlChangeOnNoteWave>,
-    pub cc_on_note_wave: Vec<ControlChangeOnNoteWave>,
-}
-
-impl Track {
-    pub fn new(timebase: isize, channel: isize) -> Self {
-        let channel = if channel < 0 { 0 } else if channel > 15 { 15 } else { channel };
-        Track {
-            timepos: 0,
-            length: timebase,
-            velocity: 100,
-            octave: 5,
-            qlen: 90,
-            timing: 0,
-            track_key: 0,
-            port: 0,
-            tie_mode: TieMode::Port,
-            tie_value: 0,
-            v_sub: vec![0],
-            v_rand: 0,
-            q_rand: 0,
-            t_rand: 0,
-            o_rand: 0,
-            program_change: 0,
-            cc_on_time_freq: 4,
-            v_on_time_start: -1,
-            v_on_time: None,
-            v_on_note_index: 0,
-            v_on_note_is_cycle: false,
-            v_on_note: None,
-            q_on_note_index: 0,
-            q_on_note_is_cycle: false,
-            q_on_note: None,
-            t_on_note_index: 0,
-            t_on_note: None,
-            t_on_note_is_cycle: false,
-            o_on_note_index: 0,
-            o_on_note: None,
-            o_on_note_is_cycle: false,
-            l_on_note_index: 0,
-            l_on_note: None,
-            l_on_note_is_cycle: false,
-            channel,
-            events: vec![],
-            tie_notes: vec![],
-            bend_range: -1,
-            cc_on_note: vec![],
-            cc_on_note_wave: vec![],
-        }
-    }
-
-    pub fn split_note_off(&self) -> Vec<Event> {
-        let mut events: Vec<Event> = vec![];
-        for i in 0..self.events.len() {
-            let e = &self.events[i];
-            match e.etype {
-                EventType::NoteOn => {
-                    events.push(e.clone());
-                    let mut noteoff = e.clone();
-                    noteoff.etype = EventType::NoteOff;
-                    noteoff.time = e.time + e.v2;
-                    events.push(noteoff);
-                },
-                _ => {
-                    events.push(e.clone());
-                }
-            }
-        }
-        events
-    }
-
-    pub fn normalize(&mut self) {
-        let events: Vec<Event> = self.split_note_off();
-        self.events = events;
-    }
-    pub fn events_sort(&mut self) {
-        // sort_byなら要素の順序は保持される
-        self.events.sort_by(|a, b| a.time.cmp(&b.time));
-    }
-    pub fn play_from(&mut self, timepos: isize) {
-        let mut events: Vec<Event> = vec![];
-        let mut cc_values: Vec<isize> = vec![];
-        let mut voice: isize = -1;
-        let mut ch: isize = 0;
-        for _ in 0..128 { cc_values.push(-1); }
-        for e in self.events.iter() {
-            match e.etype {
-                EventType::Meta | EventType::SysEx => {
-                    let mut e2 = e.clone();
-                    e2.time -= timepos;
-                    if e2.time < 0 { e2.time = 0; }
-                    events.push(e2);
-                },
-                EventType::NoteOn => {
-                    let mut e2 = e.clone();
-                    e2.time -= timepos;
-                    if e2.time < 0 { continue; }
-                    events.push(e2);
-                },
-                EventType::Voice => {
-                    let mut e2 = e.clone();
-                    e2.time -= timepos;
-                    if e2.time < 0 {
-                        voice = e2.v1;
-                        ch = e2.channel;
-                        continue;
-                    }
-                    events.push(e2);
-                },
-                EventType::ControllChange => {
-                    let mut e2 = e.clone();
-                    e2.time -= timepos;
-                    if e2.time < 0 {
-                        cc_values[e2.v1 as usize] = e2.v2;
-                        ch = e2.channel;
-                        continue;
-                    }
-                    events.push(e2);
-                },
-                EventType::NoteOff => {},
-                EventType::PitchBend => {}, // TODO: #8
-                EventType::PitchBendRange => {}, // TODO: #8
-                EventType::DirectSMF => {},
-            }
-        }
-        // add cc
-        for no in 0..128 {
-            if cc_values[no] < 0 { continue; }
-            events.push(Event::cc(0, ch, no as isize, cc_values[no as usize]));
-        }
-        // voice
-        if voice >= 0 {
-            events.push(Event::voice(0, ch, voice));
-        }
-        self.events = events;
-    }
-    pub fn calc_v_on_time(&mut self, def: isize) -> isize {
-        let start_time = self.v_on_time_start;
-        let cur_time = self.timepos - start_time;
-        let mut result = isize::MIN;
-        // on_time?
-        let ia = match &self.v_on_time {
-            None => return def,
-            Some(pia) => pia.clone()
-        };
-        let mut area_time = 0;
-        for i in 0..ia.len() / 3 {
-            let low = ia[i*3+0];
-            let high = ia[i*3+1];
-            let len = ia[i*3+2];
-            let area_time_to = area_time + len;
-            if area_time <= cur_time && cur_time < area_time_to {
-                let v = (high - low) as f32 * ((cur_time - area_time) as f32 / len as f32) + low as f32;
-                result = v as isize;
-            }
-            area_time = area_time_to;
-        }
-        // over ?
-        if area_time <= cur_time {
-            self.v_on_time = None;
-            self.v_on_time_start = -1;
-        }
-        if result == isize::MIN { result = def; }
-        result
-    }
-    pub fn calc_v_on_note(&mut self, def: isize) -> isize {
-        // on_note?
-        let ia = match &self.v_on_note {
-            None => return def,
-            Some(pia) => pia.clone()
-        };
-        if ia.len() == 0 { return def; }
-        if self.v_on_note_index >= ia.len() as isize {
-            if self.v_on_note_is_cycle {
-                self.v_on_note_index = 0;
-            } else {
-                self.v_on_note = None;
-                self.v_on_note_index = 0;
-                return def;
-            }
-        }
-        let v = ia[(self.v_on_note_index as usize) % ia.len()];
-        self.velocity = v;
-        self.v_on_note_index += 1;
-        return v;
-    }
-    pub fn calc_t_on_note(&mut self, def: isize) -> isize {
-        // on_note?
-        let ia = match &self.t_on_note {
-            None => return def,
-            Some(pia) => pia.clone()
-        };
-        if ia.len() == 0 { return def; }
-        if self.t_on_note_index >= ia.len() as isize {
-            if self.t_on_note_is_cycle {
-                self.t_on_note_index = 0;
-            } else {
-                self.t_on_note = None;
-                self.t_on_note_index = 0;
-                return def;
-            }
-        }
-        let t = ia[(self.t_on_note_index as usize) % ia.len()];
-        self.timing = t;
-        self.t_on_note_index += 1;
-        return t;
-    }
-    pub fn calc_qlen_on_note(&mut self, def: isize) -> isize {
-        // on_note?
-        let ia = match &self.q_on_note {
-            None => return def,
-            Some(pia) => pia.clone()
-        };
-        if ia.len() == 0 { return def; }
-        if self.q_on_note_index >= ia.len() as isize {
-            if self.q_on_note_is_cycle {
-                self.q_on_note_index = 0;
-            } else {
-                self.q_on_note = None;
-                self.q_on_note_index = 0;
-                return def;
-            }
-        }
-        let qlen = ia[(self.q_on_note_index as usize) % ia.len()];
-        self.qlen = qlen;
-        self.q_on_note_index += 1;
-        return qlen;
-    }
-    pub fn calc_o_on_note(&mut self, def: isize) -> isize {
-        // on_note?
-        let ia = match &self.o_on_note {
-            None => return def,
-            Some(pia) => pia.clone()
-        };
-        if ia.len() == 0 {
-            self.o_on_note = None;
-            return def;
-        }
-        if self.o_on_note_index >= ia.len() as isize {
-            if self.o_on_note_is_cycle {
-                self.o_on_note_index = 0;
-            } else {
-                self.o_on_note = None;
-                self.o_on_note_index = 0;
-                return def;
-            }
-        }
-        let o = ia[(self.o_on_note_index as usize) % ia.len()];
-        self.octave = o;
-        self.o_on_note_index += 1;
-        return o;
-    }
-    pub fn calc_l_on_note(&mut self, def: isize) -> isize {
-        // on_note?
-        let ia = match &self.l_on_note {
-            None => return def,
-            Some(pia) => pia.clone()
-        };
-        if ia.len() == 0 {
-            self.l_on_note = None;
-            return def;
-        }
-        if self.l_on_note_index >= ia.len() as isize {
-            if self.l_on_note_is_cycle {
-                self.l_on_note_index = 0;
-            } else {
-                self.l_on_note = None;
-                self.l_on_note_index = 0;
-                return def;
-            }
-        }
-        let l = ia[(self.l_on_note_index as usize) % ia.len()];
-        self.l_on_note_index += 1;
-        return l;
-    }
-    pub fn write_cc_on_time(&mut self, cc_no: isize, ia: Vec<isize>) {
-        let freq = self.cc_on_time_freq;
-        for i in 0..ia.len() / 3 {
-            let low = ia[i*3+0];
-            let high = ia[i*3+1];
-            let len = ia[i*3+2];
-            // println!("CC.T={},{},{}", low, high, len);
-            for j in 0..len {
-                if (j % freq) == 0 {
-                    let v = (high - low) as f32 * (j as f32 / len as f32) + low as f32;
-                    let v = value_range(0, v as isize, 127);
-                    let e = Event::cc(self.timepos + j, self.channel, cc_no, v);
-                    self.events.push(e);
-                }
-            }
-        }
-    }
-    pub fn write_pb_on_time(&mut self, is_big: isize, ia: Vec<isize>, timebase: isize) {
-        let freq = timebase / 32;
-        for i in 0..ia.len() / 3 {
-            let mut low = ia[i*3+0];
-            let mut high = ia[i*3+1];
-            if is_big == 0 { // small
-                low = low * 128;
-                high = high * 128;
-            } else { // big
-                low += 8192;
-                high += 8192;
-            }
-            // println!("@@@PB.T={},{}", low,high);
-            let len = ia[i*3+2];
-            for j in 0..len {
-                if (j % freq) == 0 {
-                    let v = (high - low) as f32 * (j as f32 / len as f32) + low as f32;
-                    let v = value_range(0, v as isize, 0x7f7f);
-                    let e = Event::pitch_bend(self.timepos + j, self.channel, v);
-                    self.events.push(e);
-                }
-            }
-        }
-    }
-    pub fn remove_cc_on(&mut self, no: isize) {
-        self.remove_cc_on_note(no);
-        self.remove_cc_on_note_wave(no);
-    }
-    pub fn remove_cc_on_note_wave(&mut self, no: isize) {
-        if self.cc_on_note_wave.len() == 0 { return; }
-        let mut new_list: Vec<ControlChangeOnNoteWave> = vec![];
-        for cow in self.cc_on_note_wave.iter() {
-            if cow.no == no { continue; }
-            new_list.push(cow.clone());
-        }
-        self.cc_on_note_wave = new_list;
-    }
-    pub fn set_cc_on_note_wave(&mut self, no: isize, ia: Vec<isize>) {
-        self.remove_cc_on(no);
-        let cc_new = ControlChangeOnNoteWave { no, data: ia, index: 0 };
-        self.cc_on_note_wave.push(cc_new);
-    }
-    pub fn write_cc_on_note_wave(&mut self, start_pos: isize) {
-        let end_pos = self.timepos;
-        // let _wave_len = end_pos - start_pos;
-        self.timepos = start_pos;
-        for cow in self.cc_on_note_wave.clone().iter() {
-            // println!("write_cc_on_note_wave:no={}", cow.no);
-            self.write_cc_on_time(cow.no, cow.data.clone());
-        }
-        self.timepos = end_pos;
-    }
-    pub fn remove_cc_on_note(&mut self, no: isize) {
-        if self.cc_on_note.len() == 0 { return; }
-        let mut new_list: Vec<ControlChangeOnNoteWave> = vec![];
-        for cow in self.cc_on_note.iter() {
-            if cow.no == no { continue; }
-            new_list.push(cow.clone());
-        }
-        self.cc_on_note = new_list;
-    }
-    pub fn set_cc_on_note(&mut self, no: isize, ia: Vec<isize>) {
-        self.remove_cc_on(no);
-        let cc_new = ControlChangeOnNoteWave { no, data: ia, index: 0 };
-        self.cc_on_note.push(cc_new);
-    }
-    pub fn write_cc_on_note(&mut self, start_pos: isize) {
-        for it in self.cc_on_note.iter_mut() {
-            if it.data.len() <= it.index as usize {
-                continue;
-            }
-            let v = it.data[it.index as usize];
-            it.index += 1;
-            let e = Event::cc(start_pos, self.channel, it.no, v);
-            self.events.push(e);
-        }
-        self.cc_on_note.retain(|it| it.data.len() > it.index as usize);
-    }
-}
-
-#[derive(Debug)]
-pub struct Flags {
-    pub harmony_flag: bool,
-    pub harmony_time: isize,
-    pub harmony_events: Vec<Event>,
-    pub octave_once: isize,
-    pub measure_shift: isize,
-    pub break_flag: isize, // 0: none 1: break 2: continue 3: return
-    pub max_loop: isize,
-    pub function_needs_return_value: bool,
-}
-impl Flags {
-    pub fn new() -> Self {
-        Flags {
-            harmony_flag: false,
-            harmony_time: 0,
-            harmony_events: vec![],
-            octave_once: 0,
-            measure_shift: 0,
-            break_flag: 0,
-            max_loop: 10000,
-            function_needs_return_value: false,
-        }
-    }
-}
-
-#[derive(Debug)]
-pub enum SFunctionType {
-    System,
-    User,
-}
-
-#[derive(Debug)]
-pub struct SFunction {
-    pub name: String,
-    pub tokens: Tokens,
-    pub lineno: isize,
-    pub func_id: usize,
-    pub arg_names: Vec<String>,
-    pub arg_types: Vec<char>, // S: string, I: int, A: array
-    pub arg_def_values: Vec<SValue>,
-    pub function_type: SFunctionType,
-}
-
-impl SFunction {
-    pub fn new(name: &str, tokens: Tokens, func_id: usize, lineno: isize) -> Self {
-        Self {
-            name: name.to_string(),
-            tokens,
-            lineno,
-            func_id,
-            arg_names: vec![],
-            arg_types: vec![],
-            arg_def_values: vec![],
-            function_type: SFunctionType::User,
-        }
-    }
-    pub fn new_system(name: &str, func_id: usize, arg_types: &'static str) -> Self {
-        Self {
-            name: name.to_string(),
-            tokens: vec![],
-            lineno: 0,
-            func_id,
-            arg_names: vec![],
-            arg_types: arg_types.chars().collect(),
-            arg_def_values: vec![],
-            function_type: SFunctionType::System,
-        }
-    }
-}
 
 /// Song
 #[derive(Debug)]
